refactor country's page
